Add org workflow (#613)
* Update styling for Complete component

* Move workflow to be a modal for the TagCreator page
